Restructuring provisioning scripts for better organization...
I also have fixed a few bugs. consolas will now be sure to remove the temp dir that it creates when downloading and extracting the fonts it needs. Also, the provision script is outputting a lot less of stdout to the console. stderr is still outputting to the console, and they should both be outputting to the log.
Thanks again, @xcjs!