Working on fixing up installation script...
Moved packages into an array instead of a text file. When packages are installed, the script first checks to see if there is a package script to run, and if not, it will just try to install the package with apt. I'm also restructuring the files and directories a bit.